requirements.txt
fastapi
uvicorn
python-dotenv
pypinyin
supabase
python-multipart